Move Functionality in File and Hadoop
