+++ spark/src/main/scala/geotrellis/spark/io/file/FileLayerReader.scala
@@ -41,6 +41,7 @@ class FileLayerReader(
 )(implicit sc: SparkContext) extends FilteringLayerReader[LayerId] with LazyLogging {
 
   val defaultNumPartitions = sc.defaultParallelism
+  def sparkContext: SparkContext = sc
 
   def read[
     K: AvroRecordCodec: Boundable: JsonFormat: ClassTag,
