+++ spark/src/main/scala/geotrellis/spark/io/hadoop/HadoopLayerUpdater.scala
@@ -39,85 +39,28 @@ class HadoopLayerUpdater(
   layerCopier: HadoopLayerCopier
 ) extends LayerUpdater[LayerId] with LazyLogging {
 
-  protected def _overwrite[
+  implicit val sc = layerReader.sparkContext
+
+  def update[
     K: AvroRecordCodec: Boundable: JsonFormat: ClassTag,
     V: AvroRecordCodec: ClassTag,
     M: JsonFormat: GetComponent[?, Bounds[K]]: Mergable
-  ](
-    id: LayerId,
-    rdd: RDD[(K, V)] with Metadata[M],
-    keyBounds: KeyBounds[K]
-  ): Unit = {
-    _update(id, rdd, keyBounds, None)
-  }
+  ](id: LayerId, rdd: RDD[(K, V)] with Metadata[M], mergeFunc: (V, V) => V): Unit =
+    layerWriter.update(id, rdd, mergeFunc)
 
-  protected def _update[
+  def update[
     K: AvroRecordCodec: Boundable: JsonFormat: ClassTag,
     V: AvroRecordCodec: ClassTag,
     M: JsonFormat: GetComponent[?, Bounds[K]]: Mergable
-  ](
-    id: LayerId,
-    rdd: RDD[(K, V)] with Metadata[M],
-    keyBounds: KeyBounds[K],
-    mergeFunc: (V, V) => V
-  ): Unit = {
-    _update(id, rdd, keyBounds, Some(mergeFunc))
-  }
+  ](id: LayerId, rdd: RDD[(K, V)] with Metadata[M]): Unit =
+    layerWriter.update(id, rdd)
 
-  def _update[
+  def overwrite[
     K: AvroRecordCodec: Boundable: JsonFormat: ClassTag,
     V: AvroRecordCodec: ClassTag,
     M: JsonFormat: GetComponent[?, Bounds[K]]: Mergable
-  ](
-    id: LayerId,
-    rdd: RDD[(K, V)] with Metadata[M],
-    keyBounds: KeyBounds[K],
-    mergeFunc: Option[(V, V) => V]
-  ): Unit = {
-    if (!attributeStore.layerExists(id)) throw new LayerNotFoundError(id)
-    val LayerAttributes(header, metadata, keyIndex, writerSchema) = try {
-      attributeStore.readLayerAttributes[HadoopLayerHeader,M, K](id)
-    } catch {
-      case e: AttributeNotFoundError => throw new LayerUpdateError(id).initCause(e)
-    }
-
-    if (!(keyIndex.keyBounds contains keyBounds))
-      throw new LayerOutOfKeyBoundsError(id, keyIndex.keyBounds)
-
-    logger.warn(s"MapFiles cannot be updated, so this requires rewriting the entire layer.")
-
-    val entireLayer = layerReader.read[K, V, M](id)
-
-    val updatedMetadata: M =
-      metadata.merge(rdd.metadata)
-
-    val fn = mergeFunc match {
-      case Some(fn) => fn
-      case None => { (v1: V, v2: V) => v2 }
-    }
-
-    val updatedRdd: RDD[(K, V)] =
-      entireLayer
-        .fullOuterJoin(rdd)
-        .flatMapValues {
-          case (Some(layerTile), Some(updateTile)) => Some(fn(layerTile, updateTile))
-          case (Some(layerTile), _) => Some(layerTile)
-          case (_, Some(updateTile)) => Some(updateTile)
-          case _ => None
-        }
-
-    val updated = ContextRDD(updatedRdd, updatedMetadata)
-
-    val tmpId = id.createTemporaryId
-    logger.info(s"Saving updated RDD to temporary id $tmpId")
-    layerWriter.write(tmpId, updated, keyIndex)
-    logger.info(s"Deleting layer $id")
-    layerDeleter.delete(id)
-    logger.info(s"Copying in $tmpId to $id")
-    layerCopier.copy[K, V, M](tmpId, id)
-    logger.info(s"Deleting temporary layer at $tmpId")
-    layerDeleter.delete(tmpId)
-  }
+  ](id: LayerId, rdd: RDD[(K, V)] with Metadata[M]): Unit =
+    layerWriter.overwrite(id, rdd)
 }
 
 object HadoopLayerUpdater {
